Isolate AssetsFilterRow into an independent component
