+++ mobile-app/app/screens/AppNavigator/screens/Portfolio/components/AssetsFilterRow.test.tsx
@@ -1,0 +1,23 @@
+import { render } from '@testing-library/react-native'
+import { AssetsFilterRow } from './AssetsFilterRow'
+
+jest.mock('@shared-contexts/ThemeProvider')
+
+describe('Portfolio Filter Buttons', () => {
+  it('should match snapshot for Assets filter Buttons component', async () => {
+    const setActiveButtonGroup = jest.fn()
+    const onButtonPress = jest.fn()
+    const setTabButtonLabel = jest.fn()
+    const activeButtonGroup = 'ALL TOKENS'
+
+    const rendered = render(
+      <AssetsFilterRow
+        setActiveButtonGroup={setActiveButtonGroup}
+        onButtonGroupPress={onButtonPress}
+        activeButtonGroup={activeButtonGroup}
+        setTabButtonLabel={setTabButtonLabel}
+      />
+    )
+    expect(rendered.toJSON()).toMatchSnapshot()
+  })
+})
